uri: rename uri.* to uri_parser.*
In the next commit we implement new `struct uri` and
rename current `struct uri` to `struct uri_raw` which
will be located in uri_parser.* files. Since git does
not understand that we renamed file, if in the same
commit we create file with the same name that file had
before renaming, we will perform the renaming in a
separate commit.

Part of #5928
