+++ test/unit/CMakeLists.txt
@@ -24,8 +24,8 @@ add_executable(heap_iterator.test heap_iterator.c)
 target_link_libraries(heap_iterator.test unit)
 add_executable(stailq.test stailq.c)
 target_link_libraries(stailq.test unit)
-add_executable(uri.test uri.c unit.c)
-target_link_libraries(uri.test uri unit)
+add_executable(uri_parser.test uri_parser.c unit.c)
+target_link_libraries(uri_parser.test uri unit)
 add_executable(queue.test queue.c)
 add_executable(mhash.test mhash.c)
 target_link_libraries(mhash.test unit)
